modified CSS | commented code | removed redundant styles
